refactor: rewrite http-server to class
